dom/base/test/test_websocket5.html

<!DOCTYPE HTML>
<html>
<head>
  <meta http-equiv="Content-Type" content="text/html;charset=UTF-8"></meta>
  <title>WebSocket test</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="websocket_helpers.js"></script>
  <script type="text/javascript" src="websocket_tests.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body onload="testWebSocket()">
<script class="testbody" type="text/javascript">

var tests = [
  test41, // HSTS
  test42, // non-char utf-8 sequences
  test43, // Test setting binaryType attribute
  test44, // Test sending/receving binary ArrayBuffer
  test45, // Test sending/receving binary Blob
  test46, // Test that we don't dispatch incoming msgs once in CLOSING state
  test47, // Make sure onerror/onclose aren't called during close()
  test48, // see bug 1227136 - client calls close() from onopen() and waits
          // until WebSocketChannel::mSocketIn is nulled out on socket thread
];

function testWebSocket() {
  doTest();
}

SimpleTest.requestFlakyTimeout("The web socket tests are really fragile, but avoiding timeouts might be hard, since it's testing stuff on the network. " +
                               "Expect all sorts of flakiness in this test...");
SimpleTest.waitForExplicitFinish();

</script>

<div id="feedback">
</div>

</body>
</html>
